Implemented check login functionality . close #24

+++ src/main/java/dians_project/mapedonija/controller/ShopController.java
@@ -1,8 +1,6 @@
 package dians_project.mapedonija.controller;
 
-import com.google.cloud.firestore.FieldValue;
-import com.google.protobuf.Timestamp;
-import com.google.protobuf.TimestampOrBuilder;
+
 import dians_project.mapedonija.model.DummyShop;
 import dians_project.mapedonija.model.Review;
 import dians_project.mapedonija.model.Shop;
@@ -10,6 +8,8 @@ import dians_project.mapedonija.service.ShopService;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ExecutionException;
@@ -25,8 +25,12 @@ public class ShopController {
     }
 
     @PostMapping("/create")
-    public String createShop(@RequestBody Shop shop) throws InterruptedException, ExecutionException {
-        return shopService.createShop(shop);
+    public String createShop(@RequestBody Shop shop, HttpServletRequest request) throws InterruptedException, ExecutionException {
+        if (LoggedInUserCheck.getInstance().check(request.getHeader("loginToken"))) {
+            return shopService.createShop(shop);
+        } else {
+            return "User timed out";
+        }
     }
 
     @GetMapping("/all")
@@ -50,15 +54,23 @@ public class ShopController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public String deleteShop(@PathVariable String id) {
-        return shopService.deleteShop(id);
+    public String deleteShop(@PathVariable String id, HttpServletRequest request) {
+        if (LoggedInUserCheck.getInstance().check(request.getHeader("loginToken"))) {
+            return shopService.deleteShop(id);
+        } else {
+            return "User timed out";
+        }
     }
 
     @PutMapping("/update/{id}")
-    public HttpStatus updateShop(@PathVariable String id, @RequestBody Map<String, Object> shop) {
-        shop.remove("id");
-        shop.remove("reviewList");
-        return shopService.updateShop(shop, id);
+    public HttpStatus updateShop(@PathVariable String id, @RequestBody Map<String, Object> shop, HttpServletRequest request) {
+        if (LoggedInUserCheck.getInstance().check(request.getHeader("loginToken"))) {
+            shop.remove("id");
+            shop.remove("reviewList");
+            return shopService.updateShop(shop, id);
+        } else {
+            return HttpStatus.UNAUTHORIZED;
+        }
     }
 
     @GetMapping("/{id}/reviews")
@@ -66,14 +78,22 @@ public class ShopController {
         return shopService.reviewList(id);
     }
 
-    @PostMapping(value = "/{id}/add-review")    //treba da se prati shopId, i plus: username, comment, grade vo objekt
-    public String addShopReview(@PathVariable String id, @RequestBody Review review) throws ExecutionException, InterruptedException{
-        return shopService.addReviews(id,review);
+    @PostMapping(value = "/{id}/add-review")
+    public String addShopReview(@PathVariable String id, @RequestBody Review review, HttpServletRequest request) throws ExecutionException, InterruptedException{
+        if (LoggedInUserCheck.getInstance().check(request.getHeader("loginToken"))) {
+            return shopService.addReviews(id,review);
+        } else {
+            return "User timed out";
+        }
     }
 
     @DeleteMapping("/{id}/delete-review")
-    public List<Review> deleteShopReview(@PathVariable String id, @RequestParam int reviewId)  throws ExecutionException, InterruptedException{
-        return shopService.deleteReview(id,reviewId);
+    public List<Review> deleteShopReview(@PathVariable String id, @RequestParam int reviewId, HttpServletRequest request)  throws ExecutionException, InterruptedException{
+        if (LoggedInUserCheck.getInstance().check(request.getHeader("loginToken"))) {
+            return shopService.deleteReview(id,reviewId);
+        } else {
+            return new ArrayList<>();
+        }
     }
 
 }
